+++ NEWS.md
@@ -18,6 +18,7 @@ Red Panda C++ Version 3.5
   - change: Remove "parse system headers" and "parse local headers" options.
   - fix: Can't get clang's version.
   - change: Show completion popup only if it's not empty.
+  - change: Reformat code don't clear breakpoints and bookmarks.
 
 Red Panda C++ Version 3.4
   - enhancement: Auto hide option "Auto clear parsed symbols when editor hidden" if "editors share one parser" is unchecked.